requirements.txt
django-1.4
django-rest-framework
